Changes to how the the list of screens is given to Remote Control, attempts to fix some cases that resulted in falling back to Legacy rendering. Attempted lazy fix for Soft Reconnect crashing. Updated Files tab to use easier to read file sizes and dates. Removed the bottom textbox from Files and Registry. Changed Process Memory from int to ulong.

+++ Modules/Processes/ProcessValue.cs
@@ -6,14 +6,14 @@ namespace KLC_Finch.Modules {
         public int PID { get; private set; }
         public string DisplayName { get; private set; }
         public string UserName { get; private set; }
-        public int Memory { get; private set; } //Normally string
+        public ulong Memory { get; private set; } //Normally string
         public int CPU { get; private set; } //Normally string
 
         public ProcessValue(dynamic p) {
             PID = (int)p["PID"];
             DisplayName = (string)p["DisplayName"];
             UserName = (string)p["UserName"];
-            Memory = int.Parse((string)p["Memory"]);
+            Memory = ulong.Parse((string)p["Memory"]);
             CPU = (int)Math.Ceiling(double.Parse((string)p["CPU"]));
         }
